+++ ui/vite.lib.config.ts
@@ -28,6 +28,8 @@ export default defineConfig({
       external: [
         'react',
         'react-dom',
+        'react/jsx-runtime',
+        'react/jsx-dev-runtime',
         'react-router-dom',
         '@mui/material',
         '@mui/icons-material',